Comments about Ubuntu20.04 need for Setonix have been added to recipes
"Recent" versions of OpenFOAM build fine with Ubuntu20.04, so just a minor comment was added to those.
"Old" versions of OpenFOAM do not even build with Ubuntu20.04 with standard procedure. So, for those
old versions, several comments were added.
Main comment is the warning that those recipes have problems building with ubuntu20.04. So those versions
will not be able to run on Setonix unless those problem are addressed. But, as versions are too old, Pawsey
will not put effort in performing this fixing, and the effort is left to interested users.
Also, we comment that old images have been removed from the main repositories and moved to "legacy"
repositories in quay.io.